+++ CHANGELOG.md
@@ -7,9 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased](https://github.com/nuwave/lighthouse/compare/v4.3.0...master)
 
+### Added
+
+- Allow overwriting the name of Enum types created through `LaravelEnumType` https://github.com/nuwave/lighthouse/pull/968
+
 ### Changed
 
 - Allow additional route configurations `prefix` and `domain` https://github.com/nuwave/lighthouse/pull/951
+- Enable schema cache only when `APP_ENV` != 'local' https://github.com/nuwave/lighthouse/pull/957
+- Improve performance of `TypeRegistry` resolving fields https://github.com/nuwave/lighthouse/pull/961
 
 ### Deprecated